app/aitools/pyproject.toml
[tool.poetry]
name = "aitools"
version = "0.1.0"
description = ""
authors = ["Vikrant Singh <vikrant@vapusdata.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
pydantic = "^2.10.5"
presidio-analyzer = "^2.2.357"
presidio-anonymizer = "^2.2.357"
grpcio = "^1.69.0"
grpcio-tools = "^1.69.0"
pylint = "^3.3.3"
sentence-transformers = "^3.3.1"
pyyaml = "^6.0.2"
grpc-interceptor = "^0.15.4"
bandit = "^1.8.2"
pyjwt = "^2.10.1"
pydantic-settings = "^2.7.1"
python-dotenv = "^1.0.1"
pyinstaller = "^6.11.1"
buf = "^1.0.0"
spacy = "^3.8.4"
cloud-sql-python-connector = "^1.16.0"
sqlalchemy = "^2.0.37"
psycopg2 = "^2.9.10"
rake-nltk = "^1.0.6"
keybert = "^0.9.0"
pytextrank = "^3.3.0"
nltk = "^3.9.1"
stopwords = "^1.0.1"
loguru = "^0.7.3"
gensim = "^4.3.3"
sumy = "^0.11.0"
psycopg2-binary = "^2.9.10"



[tool.poetry.group.dev.dependencies]
pyinstaller = "^6.11.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"